organize hooks and context
